% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{processInput}
\alias{processInput}
\title{processInput}
\usage{
processInput(input_object, cancer_type, gene_naming, log_transformed = TRUE)
}
\arguments{
\item{input_object}{dataframe containing gene expression data, columns are samples, rows are genes}

\item{cancer_type}{the 3 or 4 letter code of the cancer type in the submitted samples. Can only be 1 type.}

\item{gene_naming}{which type of naming scheme are used to specify the genes? set to 'name' for names, and 'ensg' for ENSG}

\item{log_transformed}{Optional Parameter, the default value is TRUE. If the expression data needs to get log2 transformed, manually specify FALSE}
}
\value{
A list formatted object that can be passed to the "QuiescenceScore" function
}
\description{
This function takes user supplied data frames of gene expression data, and formats them to be suitable for the 'QuiescenceScore' function
}
\examples{
patient1_data_processed <- processInput(patient1_data_raw, cancer_type = "LUAD", gene_naming = "name")
patient2_data_processed <- processInput(patient2_data_raw, cancer_type = "BRCA", gene_naming = "ensg")
patient3_data_processed <- processInput(patient3_data_raw, cancer_type = "PAAD", gene_naming = "ensg", log_transformed = FALSE)
}
