% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{QuiescenceScore}
\alias{QuiescenceScore}
\title{QuiescenceScore}
\usage{
QuiescenceScore(
  input_data,
  genelist_eval = "full",
  method_eval = "zscore",
  subtype_score = FALSE,
  vis = TRUE,
  parallel = 4
)
}
\arguments{
\item{input_data}{a file generated using the 'processInput' function}

\item{genelist_eval}{specify whether the quiescence phenotype should be assessed using the full 139 genes or a core set of 18 genes. Default is "full", manually specify if the "core" set should be used}

\item{method_eval}{specify the methodology used to calculate quiescence scores from the expression data. Default mode is "zscore", manually set to "gsva" or "ssgsea" if other methods should be used.}

\item{subtype_score}{specify whether the samples should be scored according to their quiescence subtypes. Default is TRUE, manually set to FALSE if only the main quiescence score should be calculated}

\item{vis}{specify whether the results of the analysis should be visualized in a PCA plot. Default is TRUE, manually set to FALSE if PCA visualizations are not needed.}

\item{parallel}{specify the number of CPU threads that will be used in the calculation. Also impacts memory usage, with larger values requiring larger memory. If an analysis fails due to a lack of memory, lower this value}
}
\value{
a data frame that details the general and optionally sub-type specific quiescence score of samples, and the percentile value of the score compared to a TCGA data set from the same cancer.
}
\description{
This function calculates general and sub-type specific quiescence scores using a variety of methods, and also generates a percentile figure which compares the score to a TCGA dataset from the same cancer type.
}
\examples{
patient1_results <- QuiescenceScore(patient1_data_processed)
patient2_results <- QuiescenceScore(patient2_data_processed, genelist_eval = "core")
patient3_results <- QuiescenceScore(patient3_data_processed, subtype_score = TRUE, vis = FALSE)
patient4_results <- QuiescenceScore(patient4_data_processed, subtype_score = TRUE, parallel = 8)
}
